🧼  Business Users had more differences
The Business User had some interesting differences - like the EIN has
to be in all caps, and the keyPersons have to have citizenship, and
so we had to make a few changes. Nothing horrible, but we also wanted to
add in the Business User Tests, just to make sure it was all working. It
is now.