🔩  Went back to lower-case 'ein'
There is no additional value in having the field EIN vs. ein, and the
casing of the inputs could only be confusing things. So let's go back
to the lower-case 'ein', and then see what's happening. The TypeScript
tests pass fine, but the Javascript ones do not... Not exactly sure why.